breaking ions into separate files
