+++ openpnm/phases/__init__.py
@@ -81,6 +81,6 @@ from .Air import Air
 from .Water import Water
 from .Mercury import Mercury
 from .Multiphase import MultiPhase
-from .Mixtures import SalineWater
 from .GenericMixture import GenericMixture
 from . import components
+from .SalineWater import SalineWater
