+++ openpnm/phases/components/ions/Cl.py
@@ -25,11 +25,14 @@ class Cl(GenericPhase):
     --------
     >>> import openpnm as op
     >>> pn = op.network.Cubic(shape=[5, 5, 5])
-    >>> Na = op.phases.components.ions.NaIon(network=pn)
+    >>> Cl = op.phases.components.ions.Cl(network=pn)
 
     """
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
-        self['pore.molecular_weight'] = 0.0291
-        self['pore.diffusivity'] = 0.1
+        self['pore.molecular_weight'] = 0.03545  # kg/mol
+        self['pore.diffusivity'] = 2.03e-09  # m2/s
+        self['throat.diffusivity'] = 2.03e-09  # m2/s
+        self['pore.valence'] = -1
+        self['throat.valence'] = -1
