+++ openpnm/phases/components/ions/__init__.py
@@ -1,1 +1,3 @@
-from . import Na, Cl, Proton
+from .Cl import Cl
+from .Na import Na
+from .Proton import Proton
